XTTS-v2: trim a run-on tail instead of returning it
The model sometimes keeps emitting codes after the sentence rather than STOP, and
they vocode to silence: measured at 8 of 650 runs, 0.6 to 2.7 s. The rate was
gated but the audio was still handed to the caller.

generate() now measures the trailing silence and cuts it back to a natural tail
when it exceeds RUN_ON_S, before the fade so the new ending does not click. The
threshold sits in an empty band -- a natural tail runs to about 0.16 s and every
run-on measured was 0.6 s or more -- so healthy audio is untouched. Verified
across 17 languages x 2 speakers: one trim in 170 runs, the known Polish case, and
tails of 0.12 to 0.18 s whether trimmed or not.

Two things had to move with it.

_trailing_silence was in test_wer and is now in the model, since the model is what
needs it. The test imports it, so its metric test now validates shipping code
rather than a copy.

Trimming would have made the run-on gate vacuous: the check measured silence in
the returned waveform, which is exactly what is now removed. generate() reports
what it cut as last_timings["run_on_s"] and the gate reads that, so the fault is
still counted even though the caller no longer sees it. The smoke test's duration
invariant adds the same figure back, so it still asserts the audio matches the
codes generated rather than being loosened to pass.

This does not stop the model generating the codes, and does nothing for the other
run-on shape where it invents audible words past the end. WER catches that one.
